abstracted out member routes
